requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
sqlalchemy
